Renamed DogCard to DogCardView to fit with naming format of other files
